LPS-170494 Move OpenAPI Resource to a different class
